backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos - Azure SQL Server
sqlalchemy==2.0.25
pyodbc==5.0.1
aioodbc==0.5.0

# ML - BERT y embeddings
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Groq API
groq==0.4.2

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Producción
gunicorn==21.2.0
